Fix Audio Clicky feature
