+++ quantum/process_keycode/process_clicky.h
@@ -1,0 +1,7 @@
+#ifndef PROCESS_CLICKY_H
+#define PROCESS_CLICKY_H
+
+void clicky_play(void);
+bool process_clicky(uint16_t keycode, keyrecord_t *record);
+
+#endif
